requirements.txt
streamlit
openai
python-dotenv
langchain
langchain-openai
streamlit_chat